Move journey to proper file
